Obsidex Academy Landing Page
"Obsidex Academy Landing Page"  This is the landing page for Obsidex Academy, a website that offers educational resources for Mining. The page includes information about the courses and services offered by the academy, as well as a form for interested users to sign up for updates. The project is built using HTML, CSS, Bootstrap